ml-kem: implement DER document structure suggested in IETF draft spec
Let us consider the test vectors from this repository:

  https://github.com/lamps-wg/kyber-certificates/tree/624bcaa4bd9ea9e72de5b51d81ce2d90cbd7e54a

This implementation uses a seed of values [0, 1, …, 63].
Now let us generate the public and private keys.

My previous implementation generated an OCTET STRING of length 1632
for ML-KEM 512:

  $ openssl asn1parse -in ml-kem-512-private-key-generated.pem -dump | head -n6 | tail -n2
   20:d=1  hl=4 l=1632 prim: OCTET STRING
      0000 - 70 55 4f d4 36 34 4f 27-85 b1 b3 b1 ba c1 84 b6   pUO.64O'........

The IETF draft specification suggests a structure involving ASN.1
CHOICE to distinguish between seed, expanded, and both formats:

  https://lamps-wg.github.io/kyber-certificates/draft-ietf-lamps-kyber-certificates.html#name-private-key-format

This leads to four additional bytes:

  $ openssl asn1parse -in ML-KEM-512-expanded.priv -dump | head -n6 | tail -n2
   20:d=1  hl=4 l=1636 prim: OCTET STRING
      0000 - 04 82 06 60 70 55 4f d4-36 34 4f 27 85 b1 b3 b1   ...`pUO.64O'....

This implementation now introduces the CHOICE discriminator.

Recognize that the draft suggests serialization to the seed format per
default. Sadly we cannot implement this. The data stored in
EncapsulationKey does not store the seed which lead to this
encapsulation key. As a result, the expanded format is always used.

The previous text talked about changes in the serialization.
In terms of deserialization all three formats must be supported which
also I implemented.