ml-kem: minimum viable seed support for `DecapsulationKey` (#133)
- Adds a `Seed` type alias for `Array<u8, U64>`
- Adds infallible `DecapsulationKey::from_seed`
- Adds fallible `DecapsulationKey::to_seed` returning an `Option` to
handle the case that the key was initialized from the expanded form.
This involved optionally carrying the original `d` as an optional struct
member of `DecapsulationKey`.

Additional notes:
- Seed-related functionality is not gated under the `deterministic`
feature. That feature should possibly be changed to only feature-gate
deterministic encryption.
- None of the existing APIs for handling the expanded form have been
updated to disambiguate them, e.g. `from_expanded_bytes`, though they
probably should be, as well as potentially being feature gated (if we
got rid of them entirely, `to_seed` could be infallible).
- Another way to make `to_seed` infallible would be an
`ExpandedDecapsulationKey` type which holds everything but `d`, with
`DecapsulationKey` being a newtype that carries `d`.
- `ml-kem` should probably adopt `subtle` to replace the built-in
constant-time code and do constant-time comparisons of private keys, as
noted in a TODO.

Closes #53