pkcs8: remove support for anything but seeds
We can re-add support for expanded decapsulation keys and "both" the
former and seeds if it ends up shipping in the final document.

For now though, we should discourage use of anything but seeds as
wasteful and needlessly complex (it was added to the draft to sate HSM
vendors who shipped expanded keys before the final FIPS 203 document and
who refuse to fix their mistakes).