ml-kem: add `DecapsulationKey::from_expanded` (#163)
This is intended to subsume the uses of the `EncodedSizeUser` trait,
providing a way for users who insist on using the expanded key format to
be able to initialize `DecapsulationKey`s.

To deter its use, it's marked as deprecated. Many implementations have
opted not to support it at all, and in the future we can potentially
consider removing it (which would allow `to_seed` to be infallible).

After this, it should be possible to deprecate `EncodedSizeUser` as
well, and remove it in a future release.